add save on exception wrapper
