+++ nle_code_wrapper/agents/sample_factory/nethack/nethack_env.py
@@ -23,7 +23,7 @@ from nle_utils.wrappers import (
 from sample_factory.utils.utils import ensure_dir_exists, experiment_dir
 
 from nle_code_wrapper.utils.utils import get_function_by_name
-from nle_code_wrapper.wrappers import NLECodeWrapper, NoProgressFeedback
+from nle_code_wrapper.wrappers import NLECodeWrapper, NoProgressFeedback, SaveOnException
 
 NETHACK_ENVS = []
 for env_spec in gym.envs.registry.all():
@@ -121,8 +121,6 @@ def make_nethack_env(env_name, cfg, env_config, render_mode: Optional[str] = Non
         panics.append(panic_func)
 
     if cfg.code_wrapper:
-        failed_game_path = join(experiment_dir(cfg=cfg), "failed_games")
-        ensure_dir_exists(failed_game_path)
         env = NLECodeWrapper(
             env,
             strategies,
@@ -132,7 +130,6 @@ def make_nethack_env(env_name, cfg, env_config, render_mode: Optional[str] = Non
             add_direction_strategies=cfg.add_direction_strategies,
             add_more_strategy=cfg.add_more_strategy,
             gamma=cfg.gamma,
-            failed_game_path=failed_game_path,
         )
         env = NoProgressFeedback(env)
 
@@ -141,5 +138,10 @@ def make_nethack_env(env_name, cfg, env_config, render_mode: Optional[str] = Non
 
     if cfg.obs_keys:
         env = ObservationFilterWrapper(env, cfg.obs_keys)
+
+    if cfg.save_on_exception:
+        failed_game_path = join(experiment_dir(cfg=cfg), "failed_games")
+        ensure_dir_exists(failed_game_path)
+        env = SaveOnException(env, failed_game_path=failed_game_path)
 
     return env
