+++ nle_code_wrapper/cfg/cfg.py
@@ -8,6 +8,7 @@ from nle_utils.utils.utils import str2bool
 
 def add_code_wrapper_cli_args(p: ArgumentParser) -> None:
     p.add_argument("--code_wrapper", type=str2bool, default=True, help="Do we want to wrap the env with code wrapper")
+    p.add_argument("--save_on_exception", type=str2bool, default=True, help="Do we want to save demo on exception")
     p.add_argument("--strategies", type=ast.literal_eval, default=[], help="List of strategy names")
     p.add_argument(
         "--strategies_loc",
