+++ nle_code_wrapper/wrappers/__init__.py
@@ -1,4 +1,5 @@
 from nle_code_wrapper.wrappers.nle_code_wrapper import NLECodeWrapper
 from nle_code_wrapper.wrappers.no_progress_feedback import NoProgressFeedback
+from nle_code_wrapper.wrappers.save_on_exception import SaveOnException
 
-__all__ = [NoProgressFeedback, NLECodeWrapper]
+__all__ = [NoProgressFeedback, NLECodeWrapper, SaveOnException]
